move all unit tests to test folder for consistency
