+++ src/pubchem/fetch.rs
@@ -53,11 +53,3 @@ pub async fn down_all_current_sdf(p: impl AsRef<Path>) -> eyre::Result<()> {
     }
     Ok(())
 }
-
-#[cfg(test)]
-mod test {
-    // use tokio::io::AsyncWriteExt;
-
-    #[tokio::test]
-    async fn get_index() {}
-}
