+++ src/rest_api/api/compound_processing/standardize.rs
@@ -30,44 +30,3 @@ pub async fn v1_standardize(
 
     StandardizeResponse::Ok(Json(standardized_smiles))
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::rest_api::api::compound_processing::standardize::*;
-    use crate::{indexing::index_manager::IndexManager, rest_api::openapi_server::Api};
-    use poem::{handler, Route};
-    use poem_openapi::param::Query;
-
-    #[handler]
-    async fn index() -> StandardizeResponse {
-        let smiles = Json(vec![Smiles {
-            smiles: "CC=CO".to_string(), // smiles:  "CCC=O".to_string(), -answer
-        }]);
-        Api {
-            index_manager: IndexManager::new("/tmp/blah", false).unwrap(),
-        }
-        .v1_standardize(smiles, Query(Some("true".to_string())))
-        .await
-    }
-
-    #[tokio::test]
-    async fn test_poem() {
-        let app = Route::new().at("/", poem::post(index));
-        let client = poem::test::TestClient::new(app);
-
-        let resp = client.post("/").send().await;
-
-        resp.assert_status_is_ok();
-
-        let json = resp.json().await;
-        let json_value = json.value();
-        json_value
-            .array()
-            .iter()
-            .map(|value| value.object().get("smiles"))
-            .collect::<Vec<_>>()
-            .first()
-            .expect("first_value")
-            .assert_string("CCC=O");
-    }
-}
