+++ src/search/identity_search.rs
@@ -91,78 +91,3 @@ pub fn build_identity_query(
 
     query_parts.join(" AND ")
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-
-    use tantivy::{
-        doc,
-        schema::{SchemaBuilder, FAST, INDEXED, STORED, STRING},
-        IndexBuilder,
-    };
-
-    use crate::search::compound_processing::process_cpd;
-
-    #[test]
-    fn test_build_query() {
-        let descriptors: HashMap<_, _> = [("NumAtoms".to_string(), 10.0)].into_iter().collect();
-        let query = super::build_identity_query(&descriptors, &"".to_string(), &None);
-        assert_eq!(query, "NumAtoms:[10 TO 10]");
-    }
-
-    #[test]
-    fn test_identity_search() {
-        let test_smiles = "CC";
-        let (query_mol, query_fingerprint, query_descriptors) =
-            process_cpd(test_smiles, false).unwrap();
-
-        let mut builder = SchemaBuilder::new();
-
-        let smiles_field = builder.add_text_field("smiles", STRING | STORED);
-        let fingerprint_field = builder.add_bytes_field("fingerprint", FAST | STORED);
-
-        let mut doc = doc!(
-            smiles_field => test_smiles,
-            fingerprint_field => query_fingerprint.0.clone().into_vec()
-        );
-
-        for (descriptor, val) in &query_descriptors {
-            if descriptor.starts_with("Num") || descriptor.starts_with("lipinski") {
-                let current_field = builder.add_i64_field(descriptor, INDEXED | STORED);
-
-                doc.add_field_value(current_field, *val as i64);
-            } else {
-                let current_field = builder.add_f64_field(descriptor, FAST | STORED);
-
-                doc.add_field_value(current_field, *val);
-            }
-        }
-
-        let schema = builder.build();
-
-        let builder = IndexBuilder::new().schema(schema);
-        let index = builder.create_in_ram().unwrap();
-
-        let mut index_writer = index.writer_with_num_threads(1, 50 * 1024 * 1024).unwrap();
-
-        index_writer.add_document(doc).unwrap();
-        index_writer.commit().unwrap();
-
-        let reader = index.reader().unwrap();
-        let searcher = reader.searcher();
-
-        let extra_query = "".to_string();
-
-        let result = super::identity_search(
-            &searcher,
-            &query_mol,
-            &None,
-            query_fingerprint.0.as_bitslice(),
-            &query_descriptors,
-            &extra_query,
-        )
-        .unwrap();
-        assert!(result.is_some());
-    }
-}
